split reducer logic for ExploreViewContainer
