+++ superset/assets/javascripts/explore/reducers/index.js
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux';
+
+import chart from './chartReducer';
+import saveModal from './saveModalReducer';
+import explore from './exploreReducer';
+
+export default combineReducers({
+  chart,
+  saveModal,
+  explore,
+});
